test: IOHandler 테스트 코드 IOHandlerTest로 분리
